infra/lambda-test/src/requirements.txt
boto3==1.34.110
python-jose==3.3.0
requests==2.31.0
